🔨 [projects] Move class `{generator => generate}.ProjectGenerator`
